chore: split the Kubernetes-Monitor to a supervisor/scanner
- the supervisor aims to be responsible for listening on the Kubernetes-API, setting up watches, handle metadata and scheduling workers/scans
- the scanner aims to be responsible for pulling images and scanning them for their dependency graph

this commit attempts to serve two purposes:
1. organise our code a bit more so each bit has clearer responsibilities
2. prepare for splitting to two different software bits so we could scale our workers/scanners
