+++ src/supervisor/metadata-extractor.ts
