+++ src/supervisor/watchers/handlers/replication-controller.ts
