+++ src/supervisor/watchers/handlers/workload.ts
@@ -1,6 +1,6 @@
 import { IKubeObjectMetadata } from '../../types';
 import { buildWorkloadMetadata } from '../../metadata-extractor';
-import WorkloadWorker = require('../..');
+import WorkloadWorker = require('../../../scanner');
 import logger = require('../../../common/logger');
 
 export async function deleteWorkload(kubernetesMetadata: IKubeObjectMetadata, workloadName: string): Promise<void> {
