+++ test/unit/pod-watch-handler-caches.test.ts
@@ -7,12 +7,12 @@ import async = require('async');
 
 import { V1PodSpec, V1Pod } from '@kubernetes/client-node';
 import transmitterTypes = require('../../src/transmitter/types');
-import * as metadataExtractor from '../../src/scanner/metadata-extractor';
+import * as metadataExtractor from '../../src/supervisor/metadata-extractor';
 
 let pushCallCount = 0;
 sinon.stub(async, 'queue').returns({ error: () => { }, push: () => pushCallCount++ } as any);
 
-import * as pod from '../../src/scanner/watchers/handlers/pod';
+import * as pod from '../../src/supervisor/watchers/handlers/pod';
 
 tap.test('image and workload image cache', async (t) => {
   const podSpecFixture = fs.readFileSync('./test/fixtures/pod-spec.json', 'utf8');
